Remove Atom specialization and simplify code
Simplify the code further. The distinction between TurtleAtom and IDAtom was
artificial and would have made things harder to deal with later on. As it
stands, I just recently realized that even basic "turtle atoms" (e.g. +, -,
&, ^, etc) can also have replacement rules. Making sure everything is unified
only will make supporting this easier.

However, this will require another round of modification to change the parser
so that "turtle atoms" can be found in the "rule name" spot so that a rule
definition such as "+ -> -;" would be entirely legal.

Signed-off-by: joce <jocelynlegault@gmail.com>