Major language definition changes
Major language changes were required to be able to introduce context sensitive
LSystems.

As it was before, we were entering "RULE_MODE" whenever we saw the "->"
operator. This was OK for context-less system as something like this wasn't
confusing for the parser:

lsystem MyLS
{
//...
    rules
    {
        A -> AB}; // in RULE_MODE, we don't match } for end of "rules"
        // ...
    }
}

However context sensitive LSystems require us to be able to specify rule atoms
before the '->' operator, and therefore, before we can get the lexer in
RULE_MODE. So this didn't go well with the parser:

lsystem MyLS
{
//...
    rules
    {
        BB} < A > B -> ABBB; // the '}' gets read as end of rules. :-(
        // ...
    }
}

So now, every statement is preceded my a "mode definer" that allows us to set
the proper at the right time. However, it makes for a much uglier language.

I might try to make it Python-style later on, as block definition by
indentation would fix my problems.

Signed-off-by: joce <jocelynlegault@gmail.com>