Modify the turtle to use matrices to compute pos
Switch from computing the turtle position "manually" to using Matrix4x4 to
deal with the movement. The code is much simplified because of this.

To be noted, many resulting example SVG files have been modified because of
how the rounding is done. It used to be that the position was rounded after
every movement. Now only the values that are sent to be rendered are rounded;
the internal position is never rounded. The results are practically the same
and the differences are tiny (of the order of about 0.0001).

Signed-off-by: joce <jocelynlegault@gmail.com>